refactor: delete admin folder
